backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.7
pydantic==1.10.7
pydantic[email]==1.10.7
psycopg2-binary==2.9.6
python-dotenv==1.0.0
python-multipart==0.0.6 